requirements.txt
Flask==3.0.2
Flask-SocketIO==5.3.6
SQLAlchemy==2.0.29
flask-sqlalchemy==3.1.1
psycopg2-binary
requests==2.31.0
pandas==2.0.3
python-dotenv==1.0.1
alembic
asyncpg
environs==11.0.0
flask[async]
bcrypt==4.1.2
blinker==1.7.0
Bootstrap-Flask==2.4.0
cffi==1.16.0
click==8.1.7
cryptography==42.0.5
flask-form==1.0.0
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pycparser==2.22
Werkzeug==3.0.2
WTForms==3.1.2
pyjwt==2.8.0
flask_migrate
coloredlogs